# docs/swagger.yaml
basePath: /api
definitions:
  model.Complaint:
    properties:
      createdAt:
        type: string
      id:
        type: string
      jointId:
        type: string
      reason:
        type: string
      status:
        $ref: '#/definitions/model.ComplaintStatus'
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  model.ComplaintStatus:
    enum:
    - open
    - resolved
    type: string
    x-enum-varnames:
    - OpenComplaint
    - ResolvedComplaint
...
  model.ErrorResponse:
    properties:
      detail: {}
      message:
        type: string
    type: object
...
  model.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Community-driven platform built to help you discover local food hot-spots
    closer to you.
  title: Chow API
  version: "1.0"
paths:
...
  /complaints:
    get:
      consumes:
      - application/json
      description: Get all complaints with pagination. Admins only Endpoint
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Complaints retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Complaint'
                  type: array
              type: object
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all complaints
      tags:
      - complaints
  /complaints/{id}:
    get:
      consumes:
      - application/json
      description: Get a single complaint by ID
      parameters:
      - description: Complaint ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Complaint retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Complaint'
              type: object
        "404":
          description: Complaint not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get one complaint
      tags:
      - complaints
  /complaints/{id}/resolve:
    patch:
      consumes:
      - application/json
      description: Resolves a specific complaint. Admins-only Endpoint
      parameters:
      - description: Complaint ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Complaint updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Complaint'
              type: object
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: User not authorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Complaint not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resolve complaint
      tags:
      - complaints
  /complaints/me:
    get:
      consumes:
      - application/json
      description: Get all complaints made by the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User complaints retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Complaint'
                  type: array
              type: object
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user complaints
      tags:
      - complaints
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"